auth added

+++ controllers/contactsControllers.js
@@ -1,9 +1,16 @@
 import * as contactsService from "../services/contactsServices.js";
 import HttpError from "../helpers/HttpError.js";
+import { buildFilter } from "../helpers/buildFilter.js";
 
-export const getAllContacts = async (_, res, next) => {
+export const getAllContacts = async (req, res, next) => {
   try {
-    const result = await contactsService.listContacts();
+    const { _id: owner } = req.user;
+    const { page = 1, limit = 20 } = req.query;
+    const skip = (page - 1) * limit;
+    const paginationSettings = { skip, limit };
+    const filter = buildFilter(req.query, ["name", "email", "phone", "favorite"]);
+    filter.owner = owner;
+    const result = await contactsService.listContacts(filter, paginationSettings);
     res.json(result);
   } catch (error) {
     next(error);
@@ -12,7 +19,8 @@ export const getAllContacts = async (_, res, next) => {
 
 export const getOneContact = async (req, res, next) => {
   try {
-    const result = await contactsService.getContactById(req.params.id);
+    const { _id: owner } = req.user;
+    const result = await contactsService.getContactByFilter({ _id: req.params.id, owner });
     if (!result) {
       throw HttpError(404);
     }
@@ -24,8 +32,8 @@ export const getOneContact = async (req, res, next) => {
 
 export const deleteContact = async (req, res, next) => {
   try {
-    const { id } = req.params;
-    const result = await contactsService.removeContact(id);
+    const { _id: owner } = req.user;
+    const result = await contactsService.removeContact({ _id: req.params.id, owner });
     if (!result) {
       throw HttpError(404);
     }
@@ -37,7 +45,8 @@ export const deleteContact = async (req, res, next) => {
 
 export const createContact = async (req, res, next) => {
   try {
-    const result = await contactsService.addContact(req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.addContact({ ...req.body, owner });
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -46,7 +55,8 @@ export const createContact = async (req, res, next) => {
 
 export const updateContact = async (req, res, next) => {
   try {
-    const result = await contactsService.updateContact(req.params.id, req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.updateContact({ _id: req.params.id, owner }, req.body);
     if (!result) {
       throw HttpError(404);
     }
@@ -58,7 +68,14 @@ export const updateContact = async (req, res, next) => {
 
 export const updateStatusContact = async (req, res, next) => {
   try {
-    const result = await contactsService.updateContact(req.params.id, req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.updateContact(
+      {
+        _id: req.params.id,
+        owner,
+      },
+      req.body
+    );
     if (!result) {
       throw HttpError(404);
     }
